--HG--
branch : apiPermissionSupport

+++ app/protected/modules/autoresponders/jobs/AutoresponderQueueMessagesInOutboxJob.php
@@ -41,11 +41,43 @@
     class AutoresponderQueueMessagesInOutboxJob extends AutoresponderOrCampaignBaseJob
     {
         /**
+         * @see BaseJob::$loadJobQueueOnCleanupAndFallback
+         * @var bool
+         */
+        protected static $loadJobQueueOnCleanupAndFallback = true;
+
+        /**
          * @returns Translated label that describes this job type.
          */
         public static function getDisplayName()
         {
            return Zurmo::t('AutorespondersModule', 'Process autoresponder messages');
+        }
+
+        /**
+         * @see parent::resolveJobsForQueue()
+         */
+        public static function resolveJobsForQueue()
+        {
+            parent::resolveJobsForQueue();
+            $pageSize       = static::JOB_QUEUE_PAGE_SIZE;
+            $offset         = 0;
+            $timeStamp      = time();
+            do
+            {
+                $autoresponderItems = AutoresponderItem::getByProcessedAndProcessDateTime(0, $timeStamp,
+                                                            $pageSize, $offset, false);
+                $offset    = $offset + $pageSize;
+                if (is_array($autoresponderItems) && count($autoresponderItems) > 0)
+                {
+                    foreach ($autoresponderItems as $autoresponderItem)
+                    {
+                        Yii::app()->jobQueue->resolveToAddJobTypeByModelByDateTimeAttribute($autoresponderItem, 'processDateTime',
+                                                'AutoresponderQueueMessagesInOutbox');
+                    }
+                }
+            }
+            while (is_array($autoresponderItems) && count($autoresponderItems) > 0);
         }
 
         /**
@@ -61,30 +93,44 @@
 
         protected function processRun()
         {
-            $batchSize = $this->resolveBatchSize();
-            $autoresponderItemsToProcess    = AutoresponderItem::getByProcessedAndProcessDateTime(
-                                                                                        0,
-                                                                                        time(),
-                                                                                        $batchSize);
+            $batchSize                   = $this->resolveBatchSize();
+            if($batchSize != null)
+            {
+                $resolvedBatchSize = $batchSize + 1;
+            }
+            else
+            {
+                $resolvedBatchSize = null;
+            }
+            $autoresponderItemsToProcess = AutoresponderItem::getByProcessedAndProcessDateTime(0, time(),
+                $resolvedBatchSize);
             $startingMemoryUsage = memory_get_usage();
             $modelsProcessedCount = 0;
             foreach ($autoresponderItemsToProcess as $autoresponderItem)
             {
-                try
+                if($modelsProcessedCount < $batchSize || $batchSize == null)
                 {
-                    $this->processAutoresponderItemInQueue($autoresponderItem);
+                    try
+                    {
+                        $this->processAutoresponderItemInQueue($autoresponderItem);
+                    }
+                    catch (NotFoundException $e)
+                    {
+                        return $autoresponderItem->delete();
+                    }
+                    catch (NotSupportedException $e)
+                    {
+                        $this->errorMessage = $e->getMessage();
+                        return false;
+                    }
+                    $this->runGarbageCollection($autoresponderItem);
+                    $modelsProcessedCount++;
                 }
-                catch (NotFoundException $e)
+                else
                 {
-                    return $autoresponderItem->delete();
+                    Yii::app()->jobQueue->add('AutoresponderQueueMessagesInOutbox', 5);
+                    break;
                 }
-                catch (NotSupportedException $e)
-                {
-                    $this->errorMessage = $e->getMessage();
-                    return false;
-                }
-                $this->runGarbageCollection($autoresponderItem);
-                $modelsProcessedCount++;
             }
             $this->addMaxmimumProcessingCountMessage($modelsProcessedCount, $startingMemoryUsage);
             return true;
